requirements.txt
python-dotenv==1.0.1
Flask==3.0.3
asgiref==3.7.2
watchgod==0.8.2
uvicorn==0.27.0
flask_cors==4.0.1 
flask_compress==1.15 
flask_talisman==1.1.0 
flask_limiter==3.7.0 
flask_swagger_ui==4.11.1 